% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textTransformer.R
\name{textTransformer}
\alias{textTransformer}
\title{This function modifies a block of text by randomly replacing a set
number of words with a random animal.}
\usage{
textTransformer(text, num_words = 1)
}
\arguments{
\item{text}{String Character}

\item{num_words}{Integer}
}
\value{
a string with the modified text
}
\description{
This function modifies a block of text by randomly replacing a set
number of words with a random animal.
}
\examples{
text = "Your chances of being ambushed by a duck are low... but never zero!"
textTransformer(text)
}
