% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordCloud.R
\name{wordCloud}
\alias{wordCloud}
\title{Inputs a text URL to return a wordcloud in the shape of the animal created using the text of the URL based on text length and the name of the animal.}
\usage{
wordCloud(link)
}
\arguments{
\item{link}{A character vector with the URL of a website.}
}
\value{
A wordcloud image in the shape of the animal.
}
\description{
Inputs a text URL to return a wordcloud in the shape of the animal created using the text of the URL based on text length and the name of the animal.
}
\examples{
wordCloud(link)
}
