% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animalClassifier.R
\name{animalClassifier}
\alias{animalClassifier}
\title{Classify the input text as an animal
based on the proportion of unique words}
\usage{
animalClassifier(text)
}
\arguments{
\item{text}{String Character}
}
\value{
A character, the classified animal type
}
\description{
Classify the input text as an animal
based on the proportion of unique words
}
\examples{
library(tidyverse)
library(tm)
animalClassifier("I just caught a pikachu!")
}
