% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordCloud.R
\name{wordCloud}
\alias{wordCloud}
\title{Title: This function modifies a block of text by randomly replacing a set number of words with wordcloud of a random animal's name.}
\usage{
wordCloud(link)
}
\arguments{
\item{link}{String Character}
}
\value{
an image with the wordcloud
}
\description{
Title: This function modifies a block of text by randomly replacing a set number of words with wordcloud of a random animal's name.
}
\examples{
link = "https://www.britannica.com/place/Japan"
wordCloud(link)
}
